requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3
fabric2==2.2.1
Fabric3==1.14.post1
idna==2.7
invoke==1.1.0
logmatic-python==0.1.7
memory-profiler==0.52.0
paramiko==2.4.1
psutil==5.4.6
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
python-json-logger==0.1.9
six==1.11.0
